+++ src/pool.rs
@@ -7,10 +7,9 @@
 //!
 //! The [`regenerate_resource_pool`](crate::systems::regenerate_resource_pool) system will regenerate resource pools of a given type if manually added.
 
-use bevy::ecs::prelude::*;
 use bevy::utils::Duration;
+use bevy::{ecs::prelude::*, utils::HashMap};
 use core::ops::{Add, AddAssign, Div, Mul, Sub, SubAssign};
-use std::marker::PhantomData;
 use thiserror::Error;
 
 use crate::{Abilitylike, CannotUseAbility};
@@ -149,15 +148,13 @@ pub struct MaxPoolLessThanMin;
 #[derive(Component, Debug)]
 pub struct AbilityCosts<A: Abilitylike, P: Pool> {
     /// The underlying cost of each ability, stored in [`Actionlike::variants`] order.
-    cost_vec: Vec<Option<P::Quantity>>,
-    _phantom: PhantomData<A>,
+    cost_map: HashMap<A, P::Quantity>,
 }
 
 impl<A: Abilitylike, P: Pool> Clone for AbilityCosts<A, P> {
     fn clone(&self) -> Self {
         AbilityCosts {
-            cost_vec: A::variants().map(|ability| *self.get(ability)).collect(),
-            _phantom: PhantomData,
+            cost_map: self.cost_map.clone(),
         }
     }
 }
@@ -165,8 +162,7 @@ impl<A: Abilitylike, P: Pool> Clone for AbilityCosts<A, P> {
 impl<A: Abilitylike, P: Pool> Default for AbilityCosts<A, P> {
     fn default() -> Self {
         AbilityCosts {
-            cost_vec: A::variants().map(|_| None).collect(),
-            _phantom: PhantomData,
+            cost_map: HashMap::new(),
         }
     }
 }
@@ -191,9 +187,9 @@ impl<A: Abilitylike, P: Pool> AbilityCosts<A, P> {
     /// Returns `true` if the underlying resource is [`None`].
     #[inline]
     #[must_use]
-    pub fn available(&self, action: A, pool: &P) -> bool {
+    pub fn available(&self, action: &A, pool: &P) -> bool {
         if let Some(cost) = self.get(action) {
-            pool.available(*cost).is_ok()
+            pool.available(cost).is_ok()
         } else {
             true
         }
@@ -208,9 +204,9 @@ impl<A: Abilitylike, P: Pool> AbilityCosts<A, P> {
     ///
     /// Returns [`Ok(())`] if the underlying [`Pool`] can support the cost of the action.
     #[inline]
-    pub fn pay_cost(&mut self, action: A, pool: &mut P) -> Result<(), CannotUseAbility> {
+    pub fn pay_cost(&mut self, action: &A, pool: &mut P) -> Result<(), CannotUseAbility> {
         if let Some(cost) = self.get(action) {
-            pool.expend(*cost)
+            pool.expend(cost)
         } else {
             Ok(())
         }
@@ -219,15 +215,15 @@ impl<A: Abilitylike, P: Pool> AbilityCosts<A, P> {
     /// Returns a reference to the underlying [`Pool::Quantity`] cost for `action`, if set.
     #[inline]
     #[must_use]
-    pub fn get(&self, action: A) -> &Option<P::Quantity> {
-        &self.cost_vec[action.index()]
+    pub fn get(&self, action: &A) -> Option<P::Quantity> {
+        self.cost_map.get(action).copied()
     }
 
     /// Returns a mutable reference to the underlying [`Pool::Quantity`] cost for `action`, if set.
     #[inline]
     #[must_use]
-    pub fn get_mut(&mut self, action: A) -> &mut Option<P::Quantity> {
-        &mut self.cost_vec[action.index()]
+    pub fn get_mut(&mut self, action: &A) -> Option<&mut P::Quantity> {
+        self.cost_map.get_mut(action)
     }
 
     /// Sets the underlying [`Pool::Quantity`] cost for `action` to the provided value.
@@ -235,8 +231,7 @@ impl<A: Abilitylike, P: Pool> AbilityCosts<A, P> {
     /// Unless you're building a new [`AbilityCosts`] struct, you likely want to use [`Self::get_mut`].
     #[inline]
     pub fn set(&mut self, action: A, cost: P::Quantity) -> &mut Self {
-        let data = self.get_mut(action);
-        *data = Some(cost);
+        self.cost_map.insert(action, cost);
 
         self
     }
@@ -252,14 +247,14 @@ impl<A: Abilitylike, P: Pool> AbilityCosts<A, P> {
 
     /// Returns an iterator of references to the underlying non-[`None`] [`Charges`]
     #[inline]
-    pub fn iter(&self) -> impl Iterator<Item = &P::Quantity> {
-        self.cost_vec.iter().flatten()
+    pub fn iter(&self) -> impl Iterator<Item = (&A, &P::Quantity)> {
+        self.cost_map.iter()
     }
 
     /// Returns an iterator of mutable references to the underlying non-[`None`] [`Charges`]
     #[inline]
-    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut P::Quantity> {
-        self.cost_vec.iter_mut().flatten()
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = (&A, &mut P::Quantity)> {
+        self.cost_map.iter_mut()
     }
 }